backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP Client for Bright Data API
httpx==0.26.0
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# ML and NLP
scikit-learn==1.4.0
nltk==3.8.1
textblob==0.17.1

# Environment Variables
python-dotenv==1.0.0

# Utilities
pydantic==2.5.3
